refactor: split oversized market and settings classes to support components and reduce line count
